+++ docs/user/changes.rst
@@ -23,6 +23,7 @@ Release history
 - feat: device call settings with multi-line support
 - break: changed signature for :meth:`api.person_settings.personal_assistant.update <wxc_sdk.person_settings.personal_assistant.PersonalAssistantApi.update>`
 - break: AvailableAgent.numbers renamed to AvailableAgent.phone_numbers
+- feat: result for meth:`api.telephony.location.number.add <wxc_sdk.telephony.location.numbers.LocationNumbersApi.add>`
 
 1.23.0
 ------